:lipstick: separate `<Hero />` component
